+++ CHANGELOG.md
@@ -20,6 +20,8 @@ in 8.x, 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 
 ## Lab
 
+### 2026-03-10
+ * Added a workaround to bin folding bug in PackingSolver
 ### 2026-03-08
  * Improved BXF2 importer by adding an interactive placement tool
  * Updated translations